pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trade-agent"
version = "0.1.0"
description = "Indian Market Financial Analysis Agent — multi-agent decision support system"
requires-python = ">=3.11"
dependencies = [
    # LLM & Orchestration
    "langgraph>=0.6.0",
    "langgraph-checkpoint>=2.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    # MCP
    "mcp>=1.13.0",
    "fastmcp>=3.0.0",
    # Data & Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Market Data
    "yfinance>=1.2.0",
    "dhanhq>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.25.0",
    "pandas-ta>=0.3.14b",
    # Vector DB
    "qdrant-client>=1.7.0",
    # PDF Parsing
    "pdfplumber>=0.9.0",
    # News / RSS
    "feedparser>=6.0.0",
    # Infrastructure
    "redis>=5.0.0",
    "httpx>=0.24.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.0.0",
    "nest-asyncio>=1.5.0",
    # API
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    # Logging
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["libs", "agents", "mcp_servers", "api", "ingestion"]
